+++ install_files/ansible-base/roles/build-securedrop-app-code-deb-pkg/files/usr.sbin.apache2
@@ -169,6 +169,7 @@
   /var/www/securedrop/db.py r,
   /var/www/securedrop/dictionaries/adjectives.txt r,
   /var/www/securedrop/dictionaries/nouns.txt r,
+  /var/www/securedrop/execution.py r,
   /var/www/securedrop/i18n.py r,
   /var/www/securedrop/journalist.py r,
   /var/www/securedrop/journalist_app/ r,
