+++ securedrop/source_app/utils.py
@@ -1,18 +1,12 @@
-import io
-import logging
 import subprocess
 
-from datetime import datetime
 from flask import session, current_app, abort, g
-from sqlalchemy import create_engine
-from sqlalchemy.orm import sessionmaker
-from threading import Thread
 
 import typing
 
 import re
 
-from crypto_util import CryptoUtil, CryptoException
+from crypto_util import CryptoException
 from models import Source
 from passphrases import PassphraseGenerator, DicewarePassphrase
 from sdconfig import SDConfig
@@ -55,44 +49,6 @@ def generate_unique_codename(config: SDConfig) -> DicewarePassphrase:
             Source.filesystem_id == filesystem_id).all()
         if len(matching_sources) == 0:
             return passphrase
-
-
-def get_entropy_estimate() -> int:
-    with io.open('/proc/sys/kernel/random/entropy_avail') as f:
-        return int(f.read())
-
-
-def asynchronous(f):               # type: ignore
-    def wrapper(*args, **kwargs):  # type: ignore
-        thread = Thread(target=f, args=args, kwargs=kwargs)
-        thread.start()
-    return wrapper
-
-
-@asynchronous
-def async_genkey(crypto_util_: CryptoUtil,
-                 db_uri: str,
-                 filesystem_id: str,
-                 codename: DicewarePassphrase) -> None:
-    # We pass in the `crypto_util_` so we don't have to reference `current_app`
-    # here. The app might not have a pushed context during testing which would
-    # cause this asynchronous function to break.
-    crypto_util_.genkeypair(filesystem_id, codename)
-
-    # Register key generation as update to the source, so sources will
-    # filter to the top of the list in the journalist interface if a
-    # flagged source logs in and has a key generated for them. #789
-    session = sessionmaker(bind=create_engine(db_uri))()
-    try:
-        source = session.query(Source).filter(
-            Source.filesystem_id == filesystem_id).one()
-        source.last_updated = datetime.utcnow()
-        session.commit()
-    except Exception as e:
-        logging.getLogger(__name__).error(
-                "async_genkey for source (filesystem_id={}): {}"
-                .format(filesystem_id, e))
-    session.close()
 
 
 def normalize_timestamps(filesystem_id: str) -> None:
